RA-1174 - :hammer: Moving apprenticeship mapping tests under feature tests folder for GettingApprenticeshipVacancy
