+++ src/Esfa.Vacancy.Register.UnitTests/GetApprenticeshipVacancy/Api/Mappings/GivenAnApprenticeshipMapper/WhenGettingLiveVacancyWage.cs
@@ -6,7 +6,7 @@ using Moq;
 using NUnit.Framework;
 using Ploeh.AutoFixture;
 
-namespace Esfa.Vacancy.Register.UnitTests.Api.Mappings.ApprenticeshipMapper
+namespace Esfa.Vacancy.Register.UnitTests.GetApprenticeshipVacancy.Api.Mappings.GivenAnApprenticeshipMapper
 {
     [TestFixture]
     public class WhenGettingLiveVacancyWage
